requirements.txt
# Core scientific computing and data analysis
numpy>=1.25
pandas>=2.2
scipy>=1.11

# Machine learning
scikit-learn>=1.3

# Data visualization
matplotlib>=3.8
seaborn>=0.13

# Image processing
Pillow>=10.0

# Jupyter notebooks
jupyter>=1.0
ipykernel>=6.0

# Astronomy and other scientific tools
astropy>=5.0
tqdm>=4.0
shap>=0.43.0

# Streamlit Map Application Dependencies
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.11.0
geopy>=2.4.0
geojson>=3.1.0
